sy - cleaning the results tab chart/plot
