+++ Examples/qfem-0009/src/input.json
@@ -23,6 +23,24 @@
         "uqEngine": "SimCenterUQ",
         "uqType": "Sensitivity Analysis"
     },
+    "correlationMatrix": [
+        1,
+        0,
+        0,
+        0,
+        0,
+        1,
+        0,
+        0,
+        0,
+        0,
+        1,
+        0.5,
+        0,
+        0,
+        0.5,
+        1
+    ],
     "fem": {
         "dir": "{Current_Dir}/.",
         "inputFile": "{Current_Dir}/TrussModel.tcl",
