+++ SimCenterUQResultsSampling_original_saved.h
@@ -43,7 +43,6 @@ UPDATES, ENHANCEMENTS, OR MODIFICATIONS.
 #include <QtCharts/QChart>
 #include <QMessageBox>
 #include <QPushButton>
-#include <ResultsDataChart.h>
 
 
 using namespace QtCharts;
@@ -67,15 +66,15 @@ public:
     bool inputFromJSON(QJsonObject &rvObject);
 
     int processResults(QString &filenameResults, QString &filenameTab);
-    QWidget *createResultEDPWidget(QString &name, QVector<double> statistics);
+    QWidget *createResultEDPWidget(QString &name, double mean, double stdDev, double skewness, double kurtosis);
 
 signals:
 
 public slots:
    void clear(void);
-//   void onSpreadsheetCellClicked(int, int);
-//   void onSaveSpreadsheetClicked();
-//   void onSaveSpreadsheetSeparatelyClicked();
+   void onSpreadsheetCellClicked(int, int);
+   void onSaveSpreadsheetClicked();
+   void onSaveSpreadsheetSeparatelyClicked();
    // modified by padhye 08/25/2018
 
 private:
@@ -97,8 +96,6 @@ private:
    QVector<double>theStdDevs;
    QVector<double>theKurtosis;
    QVector<double>theSkewness;
-
-   ResultsDataChart * theDataTable;
 };
 
 #endif // SimCenterUQ_RESULTS_SAMPLING_H
